+++ src/index.ts
@@ -1,5 +1,4 @@
 export { Viewer, Dimensions } from "./Viewer";
 export { Element, ElementOptions } from "./Element";
 export { Space, Boundary, Coordinates } from "./Space";
-export { GraphicsLayer } from "./GraphicsLayer";
-export { FunctionalTileLayer, TileFunction, TileCoordinates } from "./FunctionalTileLayer";
+export { FunctionalTileLayer, TileFunction, TileCoordinates, GraphicsLayer } from "./layer";
