+++ src/layer.ts
@@ -68,3 +68,18 @@ export interface TileCoordinates {
 }
 
 export type TileFunction = (coordinates: TileCoordinates) => Promise<string>;
+
+/**
+ * Enum for Optimaze graphics tile layers.
+ */
+export enum GraphicsLayer {
+  BearingArea = 3,
+  RentableArea = 4,
+  GrossArea = 5,
+  Architect = 6,
+  Furniture = 7,
+  OuterWall = 8,
+  Elevators = 9,
+  Shafts = 11,
+  Atrium = 23
+}
